+++ examples/view-constructor/github-view.js
@@ -10,18 +10,12 @@ import path from 'node:path';
 const extname = path.extname;
 
 /**
- * Expose `GithubView`.
- */
-
-export default GithubView;
-
-/**
  * Custom view that fetches and renders
- * remove github templates. You could
+ * remote github templates. You could
  * render templates from a database etc.
  */
 
-class GithubView {
+export default class GithubView {
   constructor(name, options) {
     this.name = name;
     options = options || {};